# Optimize topology spreading filter