--- a/pkg/scheduler/framework/plugins/podtopologyspread/filtering_test.go
+++ b/pkg/scheduler/framework/plugins/podtopologyspread/filtering_test.go
@@ -33,6 +33,7 @@ import (
 	"k8s.io/kubernetes/pkg/scheduler/internal/parallelize"
 	schedulernodeinfo "k8s.io/kubernetes/pkg/scheduler/nodeinfo"
 	st "k8s.io/kubernetes/pkg/scheduler/testing"
+	"k8s.io/utils/pointer"
 )
 
 var cmpOpts = []cmp.Option{
@@ -87,9 +88,9 @@ func TestPreFilterState(t *testing.T) {
 				TpKeyToCriticalPaths: map[string]*criticalPaths{
 					"zone": {{"zone1", 0}, {"zone2", 0}},
 				},
-				TpPairToMatchNum: map[topologyPair]int32{
-					{key: "zone", value: "zone1"}: 0,
-					{key: "zone", value: "zone2"}: 0,
+				TpPairToMatchNum: map[topologyPair]*int32{
+					{key: "zone", value: "zone1"}: pointer.Int32Ptr(0),
+					{key: "zone", value: "zone2"}: pointer.Int32Ptr(0),
 				},
 			},
 		},
@@ -122,9 +123,9 @@ func TestPreFilterState(t *testing.T) {
 				TpKeyToCriticalPaths: map[string]*criticalPaths{
 					"zone": {{"zone2", 2}, {"zone1", 3}},
 				},
-				TpPairToMatchNum: map[topologyPair]int32{
-					{key: "zone", value: "zone1"}: 3,
-					{key: "zone", value: "zone2"}: 2,
+				TpPairToMatchNum: map[topologyPair]*int32{
+					{key: "zone", value: "zone1"}: pointer.Int32Ptr(3),
+					{key: "zone", value: "zone2"}: pointer.Int32Ptr(2),
 				},
 			},
 		},
@@ -159,10 +160,10 @@ func TestPreFilterState(t *testing.T) {
 				TpKeyToCriticalPaths: map[string]*criticalPaths{
 					"zone": {{"zone3", 0}, {"zone2", 2}},
 				},
-				TpPairToMatchNum: map[topologyPair]int32{
-					{key: "zone", value: "zone1"}: 3,
-					{key: "zone", value: "zone2"}: 2,
-					{key: "zone", value: "zone3"}: 0,
+				TpPairToMatchNum: map[topologyPair]*int32{
+					{key: "zone", value: "zone1"}: pointer.Int32Ptr(3),
+					{key: "zone", value: "zone2"}: pointer.Int32Ptr(2),
+					{key: "zone", value: "zone3"}: pointer.Int32Ptr(0),
 				},
 			},
 		},
@@ -195,9 +196,9 @@ func TestPreFilterState(t *testing.T) {
 				TpKeyToCriticalPaths: map[string]*criticalPaths{
 					"zone": {{"zone2", 1}, {"zone1", 2}},
 				},
-				TpPairToMatchNum: map[topologyPair]int32{
-					{key: "zone", value: "zone1"}: 2,
-					{key: "zone", value: "zone2"}: 1,
+				TpPairToMatchNum: map[topologyPair]*int32{
+					{key: "zone", value: "zone1"}: pointer.Int32Ptr(2),
+					{key: "zone", value: "zone2"}: pointer.Int32Ptr(1),
 				},
 			},
 		},
@@ -239,13 +240,13 @@ func TestPreFilterState(t *testing.T) {
 					"zone": {{"zone1", 3}, {"zone2", 4}},
 					"node": {{"node-x", 0}, {"node-b", 1}},
 				},
-				TpPairToMatchNum: map[topologyPair]int32{
-					{key: "zone", value: "zone1"}:  3,
-					{key: "zone", value: "zone2"}:  4,
-					{key: "node", value: "node-a"}: 2,
-					{key: "node", value: "node-b"}: 1,
-					{key: "node", value: "node-x"}: 0,
-					{key: "node", value: "node-y"}: 4,
+				TpPairToMatchNum: map[topologyPair]*int32{
+					{key: "zone", value: "zone1"}:  pointer.Int32Ptr(3),
+					{key: "zone", value: "zone2"}:  pointer.Int32Ptr(4),
+					{key: "node", value: "node-a"}: pointer.Int32Ptr(2),
+					{key: "node", value: "node-b"}: pointer.Int32Ptr(1),
+					{key: "node", value: "node-x"}: pointer.Int32Ptr(0),
+					{key: "node", value: "node-y"}: pointer.Int32Ptr(4),
 				},
 			},
 		},
@@ -288,12 +289,12 @@ func TestPreFilterState(t *testing.T) {
 					"zone": {{"zone1", 3}, {"zone2", 4}},
 					"node": {{"node-b", 1}, {"node-a", 2}},
 				},
-				TpPairToMatchNum: map[topologyPair]int32{
-					{key: "zone", value: "zone1"}:  3,
-					{key: "zone", value: "zone2"}:  4,
-					{key: "node", value: "node-a"}: 2,
-					{key: "node", value: "node-b"}: 1,
-					{key: "node", value: "node-y"}: 4,
+				TpPairToMatchNum: map[topologyPair]*int32{
+					{key: "zone", value: "zone1"}:  pointer.Int32Ptr(3),
+					{key: "zone", value: "zone2"}:  pointer.Int32Ptr(4),
+					{key: "node", value: "node-a"}: pointer.Int32Ptr(2),
+					{key: "node", value: "node-b"}: pointer.Int32Ptr(1),
+					{key: "node", value: "node-y"}: pointer.Int32Ptr(4),
 				},
 			},
 		},
@@ -329,12 +330,12 @@ func TestPreFilterState(t *testing.T) {
 					"zone": {{"zone2", 0}, {"zone1", 1}},
 					"node": {{"node-a", 0}, {"node-y", 0}},
 				},
-				TpPairToMatchNum: map[topologyPair]int32{
-					{key: "zone", value: "zone1"}:  1,
-					{key: "zone", value: "zone2"}:  0,
-					{key: "node", value: "node-a"}: 0,
-					{key: "node", value: "node-b"}: 1,
-					{key: "node", value: "node-y"}: 0,
+				TpPairToMatchNum: map[topologyPair]*int32{
+					{key: "zone", value: "zone1"}:  pointer.Int32Ptr(1),
+					{key: "zone", value: "zone2"}:  pointer.Int32Ptr(0),
+					{key: "node", value: "node-a"}: pointer.Int32Ptr(0),
+					{key: "node", value: "node-b"}: pointer.Int32Ptr(1),
+					{key: "node", value: "node-y"}: pointer.Int32Ptr(0),
 				},
 			},
 		},
@@ -375,12 +376,12 @@ func TestPreFilterState(t *testing.T) {
 					"zone": {{"zone1", 3}, {"zone2", 4}},
 					"node": {{"node-b", 0}, {"node-a", 1}},
 				},
-				TpPairToMatchNum: map[topologyPair]int32{
-					{key: "zone", value: "zone1"}:  3,
-					{key: "zone", value: "zone2"}:  4,
-					{key: "node", value: "node-a"}: 1,
-					{key: "node", value: "node-b"}: 0,
-					{key: "node", value: "node-y"}: 2,
+				TpPairToMatchNum: map[topologyPair]*int32{
+					{key: "zone", value: "zone1"}:  pointer.Int32Ptr(3),
+					{key: "zone", value: "zone2"}:  pointer.Int32Ptr(4),
+					{key: "node", value: "node-a"}: pointer.Int32Ptr(1),
+					{key: "node", value: "node-b"}: pointer.Int32Ptr(0),
+					{key: "node", value: "node-y"}: pointer.Int32Ptr(2),
 				},
 			},
 		},
@@ -423,12 +424,12 @@ func TestPreFilterState(t *testing.T) {
 					"zone": {{"zone1", 3}, {"zone2", 4}},
 					"node": {{"node-b", 1}, {"node-a", 2}},
 				},
-				TpPairToMatchNum: map[topologyPair]int32{
-					{key: "zone", value: "zone1"}:  3,
-					{key: "zone", value: "zone2"}:  4,
-					{key: "node", value: "node-a"}: 2,
-					{key: "node", value: "node-b"}: 1,
-					{key: "node", value: "node-y"}: 4,
+				TpPairToMatchNum: map[topologyPair]*int32{
+					{key: "zone", value: "zone1"}:  pointer.Int32Ptr(3),
+					{key: "zone", value: "zone2"}:  pointer.Int32Ptr(4),
+					{key: "node", value: "node-a"}: pointer.Int32Ptr(2),
+					{key: "node", value: "node-b"}: pointer.Int32Ptr(1),
+					{key: "node", value: "node-y"}: pointer.Int32Ptr(4),
 				},
 			},
 		},
@@ -460,7 +461,7 @@ func TestPreFilterState(t *testing.T) {
 					"node": newCriticalPaths(),
 					"rack": newCriticalPaths(),
 				},
-				TpPairToMatchNum: make(map[topologyPair]int32),
+				TpPairToMatchNum: make(map[topologyPair]*int32),
 			},
 		},
 		{
@@ -496,7 +497,7 @@ func TestPreFilterState(t *testing.T) {
 				TpKeyToCriticalPaths: map[string]*criticalPaths{
 					"zone": newCriticalPaths(),
 				},
-				TpPairToMatchNum: make(map[topologyPair]int32),
+				TpPairToMatchNum: make(map[topologyPair]*int32),
 			},
 		},
 		{
@@ -573,9 +574,9 @@ func TestPreFilterStateAddPod(t *testing.T) {
 				TpKeyToCriticalPaths: map[string]*criticalPaths{
 					"node": {{"node-b", 0}, {"node-a", 1}},
 				},
-				TpPairToMatchNum: map[topologyPair]int32{
-					{key: "node", value: "node-a"}: 1,
-					{key: "node", value: "node-b"}: 0,
+				TpPairToMatchNum: map[topologyPair]*int32{
+					{key: "node", value: "node-a"}: pointer.Int32Ptr(1),
+					{key: "node", value: "node-b"}: pointer.Int32Ptr(0),
 				},
 			},
 		},
@@ -598,9 +599,9 @@ func TestPreFilterStateAddPod(t *testing.T) {
 				TpKeyToCriticalPaths: map[string]*criticalPaths{
 					"node": {{"node-a", 1}, {"node-b", 1}},
 				},
-				TpPairToMatchNum: map[topologyPair]int32{
-					{key: "node", value: "node-a"}: 1,
-					{key: "node", value: "node-b"}: 1,
+				TpPairToMatchNum: map[topologyPair]*int32{
+					{key: "node", value: "node-a"}: pointer.Int32Ptr(1),
+					{key: "node", value: "node-b"}: pointer.Int32Ptr(1),
 				},
 			},
 		},
@@ -623,9 +624,9 @@ func TestPreFilterStateAddPod(t *testing.T) {
 				TpKeyToCriticalPaths: map[string]*criticalPaths{
 					"node": {{"node-a", 0}, {"node-b", 1}},
 				},
-				TpPairToMatchNum: map[topologyPair]int32{
-					{key: "node", value: "node-a"}: 0,
-					{key: "node", value: "node-b"}: 1,
+				TpPairToMatchNum: map[topologyPair]*int32{
+					{key: "node", value: "node-a"}: pointer.Int32Ptr(0),
+					{key: "node", value: "node-b"}: pointer.Int32Ptr(1),
 				},
 			},
 		},
@@ -648,9 +649,9 @@ func TestPreFilterStateAddPod(t *testing.T) {
 				TpKeyToCriticalPaths: map[string]*criticalPaths{
 					"node": {{"node-a", 0}, {"node-b", 2}},
 				},
-				TpPairToMatchNum: map[topologyPair]int32{
-					{key: "node", value: "node-a"}: 0,
-					{key: "node", value: "node-b"}: 2,
+				TpPairToMatchNum: map[topologyPair]*int32{
+					{key: "node", value: "node-a"}: pointer.Int32Ptr(0),
+					{key: "node", value: "node-b"}: pointer.Int32Ptr(2),
 				},
 			},
 		},
@@ -673,11 +674,11 @@ func TestPreFilterStateAddPod(t *testing.T) {
 					"zone": {{"zone2", 0}, {"zone1", 1}},
 					"node": {{"node-x", 0}, {"node-a", 1}},
 				},
-				TpPairToMatchNum: map[topologyPair]int32{
-					{key: "zone", value: "zone1"}:  1,
-					{key: "zone", value: "zone2"}:  0,
-					{key: "node", value: "node-a"}: 1,
-					{key: "node", value: "node-x"}: 0,
+				TpPairToMatchNum: map[topologyPair]*int32{
+					{key: "zone", value: "zone1"}:  pointer.Int32Ptr(1),
+					{key: "zone", value: "zone2"}:  pointer.Int32Ptr(0),
+					{key: "node", value: "node-a"}: pointer.Int32Ptr(1),
+					{key: "node", value: "node-x"}: pointer.Int32Ptr(0),
 				},
 			},
 		},
@@ -702,11 +703,11 @@ func TestPreFilterStateAddPod(t *testing.T) {
 					"zone": {{"zone1", 1}, {"zone2", 1}},
 					"node": {{"node-a", 1}, {"node-x", 1}},
 				},
-				TpPairToMatchNum: map[topologyPair]int32{
-					{key: "zone", value: "zone1"}:  1,
-					{key: "zone", value: "zone2"}:  1,
-					{key: "node", value: "node-a"}: 1,
-					{key: "node", value: "node-x"}: 1,
+				TpPairToMatchNum: map[topologyPair]*int32{
+					{key: "zone", value: "zone1"}:  pointer.Int32Ptr(1),
+					{key: "zone", value: "zone2"}:  pointer.Int32Ptr(1),
+					{key: "node", value: "node-a"}: pointer.Int32Ptr(1),
+					{key: "node", value: "node-x"}: pointer.Int32Ptr(1),
 				},
 			},
 		},
@@ -734,12 +735,12 @@ func TestPreFilterStateAddPod(t *testing.T) {
 					"zone": {{"zone2", 1}, {"zone1", 3}},
 					"node": {{"node-a", 1}, {"node-x", 1}},
 				},
-				TpPairToMatchNum: map[topologyPair]int32{
-					{key: "zone", value: "zone1"}:  3,
-					{key: "zone", value: "zone2"}:  1,
-					{key: "node", value: "node-a"}: 1,
-					{key: "node", value: "node-b"}: 2,
-					{key: "node", value: "node-x"}: 1,
+				TpPairToMatchNum: map[topologyPair]*int32{
+					{key: "zone", value: "zone1"}:  pointer.Int32Ptr(3),
+					{key: "zone", value: "zone2"}:  pointer.Int32Ptr(1),
+					{key: "node", value: "node-a"}: pointer.Int32Ptr(1),
+					{key: "node", value: "node-b"}: pointer.Int32Ptr(2),
+					{key: "node", value: "node-x"}: pointer.Int32Ptr(1),
 				},
 			},
 		},
@@ -774,12 +775,12 @@ func TestPreFilterStateAddPod(t *testing.T) {
 					"zone": {{"zone2", 1}, {"zone1", 2}},
 					"node": {{"node-a", 0}, {"node-b", 1}},
 				},
-				TpPairToMatchNum: map[topologyPair]int32{
-					{key: "zone", value: "zone1"}:  2,
-					{key: "zone", value: "zone2"}:  1,
-					{key: "node", value: "node-a"}: 0,
-					{key: "node", value: "node-b"}: 1,
-					{key: "node", value: "node-x"}: 2,
+				TpPairToMatchNum: map[topologyPair]*int32{
+					{key: "zone", value: "zone1"}:  pointer.Int32Ptr(2),
+					{key: "zone", value: "zone2"}:  pointer.Int32Ptr(1),
+					{key: "node", value: "node-a"}: pointer.Int32Ptr(0),
+					{key: "node", value: "node-b"}: pointer.Int32Ptr(1),
+					{key: "node", value: "node-x"}: pointer.Int32Ptr(2),
 				},
 			},
 		},
@@ -814,12 +815,12 @@ func TestPreFilterStateAddPod(t *testing.T) {
 					"zone": {{"zone1", 1}, {"zone2", 1}},
 					"node": {{"node-a", 1}, {"node-b", 1}},
 				},
-				TpPairToMatchNum: map[topologyPair]int32{
-					{key: "zone", value: "zone1"}:  1,
-					{key: "zone", value: "zone2"}:  1,
-					{key: "node", value: "node-a"}: 1,
-					{key: "node", value: "node-b"}: 1,
-					{key: "node", value: "node-x"}: 2,
+				TpPairToMatchNum: map[topologyPair]*int32{
+					{key: "zone", value: "zone1"}:  pointer.Int32Ptr(1),
+					{key: "zone", value: "zone2"}:  pointer.Int32Ptr(1),
+					{key: "node", value: "node-a"}: pointer.Int32Ptr(1),
+					{key: "node", value: "node-b"}: pointer.Int32Ptr(1),
+					{key: "node", value: "node-x"}: pointer.Int32Ptr(2),
 				},
 			},
 		},
@@ -895,9 +896,9 @@ func TestPreFilterStateRemovePod(t *testing.T) {
 				TpKeyToCriticalPaths: map[string]*criticalPaths{
 					"zone": {{"zone1", 1}, {"zone2", 1}},
 				},
-				TpPairToMatchNum: map[topologyPair]int32{
-					{key: "zone", value: "zone1"}: 1,
-					{key: "zone", value: "zone2"}: 1,
+				TpPairToMatchNum: map[topologyPair]*int32{
+					{key: "zone", value: "zone1"}: pointer.Int32Ptr(1),
+					{key: "zone", value: "zone2"}: pointer.Int32Ptr(1),
 				},
 			},
 		},
@@ -925,9 +926,9 @@ func TestPreFilterStateRemovePod(t *testing.T) {
 				TpKeyToCriticalPaths: map[string]*criticalPaths{
 					"zone": {{"zone1", 1}, {"zone2", 2}},
 				},
-				TpPairToMatchNum: map[topologyPair]int32{
-					{key: "zone", value: "zone1"}: 1,
-					{key: "zone", value: "zone2"}: 2,
+				TpPairToMatchNum: map[topologyPair]*int32{
+					{key: "zone", value: "zone1"}: pointer.Int32Ptr(1),
+					{key: "zone", value: "zone2"}: pointer.Int32Ptr(2),
 				},
 			},
 		},
@@ -956,9 +957,9 @@ func TestPreFilterStateRemovePod(t *testing.T) {
 				TpKeyToCriticalPaths: map[string]*criticalPaths{
 					"zone": {{"zone1", 2}, {"zone2", 2}},
 				},
-				TpPairToMatchNum: map[topologyPair]int32{
-					{key: "zone", value: "zone1"}: 2,
-					{key: "zone", value: "zone2"}: 2,
+				TpPairToMatchNum: map[topologyPair]*int32{
+					{key: "zone", value: "zone1"}: pointer.Int32Ptr(2),
+					{key: "zone", value: "zone2"}: pointer.Int32Ptr(2),
 				},
 			},
 		},
@@ -987,9 +988,9 @@ func TestPreFilterStateRemovePod(t *testing.T) {
 				TpKeyToCriticalPaths: map[string]*criticalPaths{
 					"zone": {{"zone1", 2}, {"zone2", 2}},
 				},
-				TpPairToMatchNum: map[topologyPair]int32{
-					{key: "zone", value: "zone1"}: 2,
-					{key: "zone", value: "zone2"}: 2,
+				TpPairToMatchNum: map[topologyPair]*int32{
+					{key: "zone", value: "zone1"}: pointer.Int32Ptr(2),
+					{key: "zone", value: "zone2"}: pointer.Int32Ptr(2),
 				},
 			},
 		},
@@ -1019,12 +1020,12 @@ func TestPreFilterStateRemovePod(t *testing.T) {
 					"zone": {{"zone2", 1}, {"zone1", 3}},
 					"node": {{"node-b", 1}, {"node-x", 1}},
 				},
-				TpPairToMatchNum: map[topologyPair]int32{
-					{key: "zone", value: "zone1"}:  3,
-					{key: "zone", value: "zone2"}:  1,
-					{key: "node", value: "node-a"}: 2,
-					{key: "node", value: "node-b"}: 1,
-					{key: "node", value: "node-x"}: 1,
+				TpPairToMatchNum: map[topologyPair]*int32{
+					{key: "zone", value: "zone1"}:  pointer.Int32Ptr(3),
+					{key: "zone", value: "zone2"}:  pointer.Int32Ptr(1),
+					{key: "node", value: "node-a"}: pointer.Int32Ptr(2),
+					{key: "node", value: "node-b"}: pointer.Int32Ptr(1),
+					{key: "node", value: "node-x"}: pointer.Int32Ptr(1),
 				},
 			},
 		},