**What type of PR is this?**

/kind feature

**What this PR does / why we need it**:

There are 2 optimizations in this PR:
- Use a map of pointers and atomic changes to keep the counter per node
- Set an initial capacity for the map

Benchmark results:

```
master

BenchmarkFilter/1000nodes/single-constraint-zone-56         	    1108	   1036523 ns/op
BenchmarkFilter/1000nodes/single-constraint-node-56         	     471	   2518077 ns/op
BenchmarkFilter/1000nodes/two-Constraints-zone-node-56      	     367	   3729778 ns/op

+ atomic

BenchmarkFilter/1000nodes/single-constraint-zone-56         	    1377	    999848 ns/op
BenchmarkFilter/1000nodes/single-constraint-node-56         	     662	   1698138 ns/op
BenchmarkFilter/1000nodes/two-Constraints-zone-node-56      	     508	   2269242 ns/op

+ map capacity

BenchmarkFilter/1000nodes/single-constraint-zone-56         	    1317	   1059039 ns/op
BenchmarkFilter/1000nodes/single-constraint-node-56         	     939	   1269569 ns/op
BenchmarkFilter/1000nodes/two-Constraints-zone-node-56      	     660	   1841336 ns/op
```

Note that cloning the state becomes more expensive:

```
master

BenchmarkFilter/1000nodes/single-constraint-zone/Clone-56   	  269688	      4880 ns/op
BenchmarkFilter/1000nodes/single-constraint-node/Clone-56   	    4214	    295056 ns/op
BenchmarkFilter/1000nodes/two-Constraints-zone-node/Clone-56         	    3637	    305096 ns/op

PR

BenchmarkFilter/1000nodes/single-constraint-zone/Clone-56   	  254336	      5195 ns/op
BenchmarkFilter/1000nodes/single-constraint-node/Clone-56   	    3175	    380806 ns/op
BenchmarkFilter/1000nodes/two-Constraints-zone-node/Clone-56         	    2961	    369687 ns/op
```

But since cloning only happens when doing preemption, and the difference is small, it's worth the change.

**Which issue(s) this PR fixes**:

Fixes #84936

**Special notes for your reviewer**:

Note that special-casing hostname topology (as done in #89487) doesn't work for Filter. We need the global minimum for each topology, thus we need to calculate all counters during PreFilter.

**Does this PR introduce a user-facing change?**:

```release-note
NONE
```