@Huang-Wei 2020-03-31
> And I'm doing some local tests on it.

I did the benchmark test and got similar result. That's cool!