@Huang-Wei 2020-03-31
@alculquicondor thanks for this PR. Please squash to 3 commits - benchmark, use *int64 pointer, and initialize map capacity.